+++ tests/test_local.py
@@ -1,7 +1,13 @@
+import os
+import sys
+
 from pathlib import Path
 
 from pytest import raises
-from quiltplus import QuiltLocal
+from quiltplus import QuiltLocal, QuiltResourceURI
+
+from .conftest import pytestmark  # NOQA F401
+from .conftest import TEST_PKG, PKG_URI, pytest
 
 
 def test_local_tmp():
@@ -36,3 +42,29 @@ def test_local_path():
 
     p6 = loc.check_path({})
     assert p6 == p5
+
+
+@pytest.mark.skipif(os.getenv("GITHUB_ACTIONS") == "true", reason="does not work in CI")
+async def test_local_diff_get():
+    qpkg = QuiltResourceURI(PKG_URI)
+    await qpkg.get()
+    loc = QuiltLocal({"package": TEST_PKG})
+    assert loc.local_registry
+    assert TEST_PKG in str(loc.local_cache())
+    diff = loc._diff()
+    assert diff
+    assert isinstance(diff, dict)
+    assert len(diff) > 0
+    for uri, stage in diff.items():
+        assert stage in ("add", "rm", "touch")
+        assert stage == "rm"
+
+@pytest.mark.skipif(sys.platform.startswith('win'), reason="tmp folder name issue")
+def test_local_diff_new():
+    PKG=TEST_PKG+"-new"
+    loc = QuiltLocal({"package": PKG})
+    assert loc.local_registry
+    assert PKG in str(loc.local_cache())
+    diff = loc._diff()
+    assert diff == {}
+    assert len(diff) == 0