+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.9.2 (2023-06-04)
+
+- QuiltLocal._diff directory vs local registry
+- QuiltPackage.diff => quilt+stage+{add,rm,touch}+URI
+- properly return list of URIs from REST methods
+- preliminary path support
+
 ## 0.9.1 (2023-06-03)
 
 - support get, put, patch operations